put ci visibility classes in ci visibility module
