+++ tests/tracer/test_writer.py
@@ -15,6 +15,7 @@ from six.moves import socketserver
 import ddtrace
 from ddtrace import config
 from ddtrace.constants import KEEP_SPANS_RATE_KEY
+from ddtrace.internal.ci_visibility.writer import CIVisibilityWriter
 from ddtrace.internal.compat import PY3
 from ddtrace.internal.compat import get_connection_response
 from ddtrace.internal.compat import httplib
@@ -22,7 +23,6 @@ from ddtrace.internal.encoding import MSGPACK_ENCODERS
 from ddtrace.internal.runtime import get_runtime_id
 from ddtrace.internal.uds import UDSHTTPConnection
 from ddtrace.internal.writer import AgentWriter
-from ddtrace.internal.writer import CIVisibilityWriter
 from ddtrace.internal.writer import LogWriter
 from ddtrace.internal.writer import Response
 from ddtrace.internal.writer import _human_size
